pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-authz"
version = "0.1.0"
description = "Runtime authorization middleware for MCP tool calls"
requires-python = ">=3.11"
dependencies = [
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
